LPS-23142 Rename files instead of copy+delete files to improve upgrade performance

+++ portal-impl/src/com/liferay/portlet/documentlibrary/store/FileSystemStore.java
@@ -292,48 +292,44 @@ public class FileSystemStore extends BaseStore {
 	public void updateFile(
 			long companyId, long repositoryId, long newRepositoryId,
 			String fileName)
-		throws SystemException {
+		throws PortalException, SystemException {
 
-		try {
-			File fileNameDir = getFileNameDir(
-				companyId, repositoryId, fileName);
-			File newFileNameDir = getFileNameDir(
-				companyId, newRepositoryId, fileName);
+		File fileNameDir = getFileNameDir(
+			companyId, repositoryId, fileName);
+		File newFileNameDir = getFileNameDir(
+			companyId, newRepositoryId, fileName);
 
-			FileUtil.copyDirectory(fileNameDir, newFileNameDir);
-
+		if (newFileNameDir.exists()) {
+			throw new DuplicateFileException(fileName);
+		}
+		else {
 			File parentFile = fileNameDir.getParentFile();
 
-			FileUtil.deltree(fileNameDir);
+			fileNameDir.renameTo(newFileNameDir);
 
 			deleteEmptyAncestors(parentFile);
-		}
-		catch (IOException ioe) {
-			throw new SystemException(ioe);
 		}
 	}
 
 	public void updateFile(
 			long companyId, long repositoryId, String fileName,
 			String newFileName)
-		throws SystemException {
+		throws PortalException, SystemException {
 
-		try {
-			File fileNameDir = getFileNameDir(
-				companyId, repositoryId, fileName);
-			File newFileNameDir = getFileNameDir(
-				companyId, repositoryId, newFileName);
+		File fileNameDir = getFileNameDir(
+			companyId, repositoryId, fileName);
+		File newFileNameDir = getFileNameDir(
+			companyId, repositoryId, newFileName);
 
-			FileUtil.copyDirectory(fileNameDir, newFileNameDir);
-
+		if (newFileNameDir.exists()) {
+			throw new DuplicateFileException(newFileName);
+		}
+		else {
 			File parentFile = fileNameDir.getParentFile();
 
-			FileUtil.deltree(fileNameDir);
+			fileNameDir.renameTo(newFileNameDir);
 
 			deleteEmptyAncestors(parentFile);
-		}
-		catch (IOException ioe) {
-			throw new SystemException(ioe);
 		}
 	}
